[feat] (all) Extract balance, utxo, and nonce methods to PaymentsUtils
